Give up includeexpr, use just gf that understands tsconfig
It's very hard to get an includeexpr for JavaScript right.

This almost works except for included files INSIDE other included files.

Because I use the current file directory, these deeper files can't
possibly know the directory in which they are in.

It's possible to just expand tsconfig aliases but this is useless too
because the point is to get a index file.

How would I go about finding an index file if I only have the current
file directory and v:fname? Nothing occurs to me so I'll just abandon
it.