Check if should use js formatter in efm-langserver asynchronously
Before that, I was using os.execute to check if the eslint configuration
was working. Based on the exit_code, I'd set the
`.resolved_capabilities.document_formatting` to `true` if it was 0.

But this froze the editor on start up/opening new projects because it is
synchronous.

Now this is done in an asynchronous way with `jobstart` with an
`on_exit` callback.