Improve dwm handling of Android Emulator windows
This patch improves how to handle those Android Emulator extra windows
(loading, emulator controls etc.).

I usually want to switch the editor and emulator window with Alt+Tab or
Alt+Shift+Tab but these windows get in the way.

This fixes it by changing the `focusstack` function. It won't focus one
of these windows i.e. windows whose `c->name` is literally "Emulator"
(that's the value of their `_NET_WM_NAME`).

Also we don't want them to appear in the statusbar, so I handled that in
awesomebar logic as well.

One last thing to fix is what happens when we move the window with the
emulator, because I want these other extra windows to always follow the
emulator itself. This is done by patching the `tag` function. We
basically do the same thing that we do with the emulator to every window
whose `_NET_WM_NAME` is "Emulator".