Add command to open Gitlab merge request of a source branch
If there is more than one merge requests containing this source branch,
we should be to pick using an input. If there is only one, open it
automatically.

This depends on curl and firefox to work. Also, we should have a way to
safely get the private token, I used pass here.