Fix transient window being lowered when focused
A transient window would be lowered after getting focus because in the
unfocus function we lower these windows if the client being unfocused is
the window leader.

This is not what we want, obviously.

The apparent fix seems to be to do this logic not in the unfocus
function, but in the restack function.

If a window is transient, we just raise it into view and be don't
restacking. But if the window is not transient (therefore, potentially
is a leader window), we need to lower all transient windows that are not
their transient windows and raise only the transient windows which have
the currently selected client as leader.