Fix getting volume level for dwm statusbar in Fedora
The issue was that nothing was showing in the line that contained
'Mono', for some reason.

Now, we get the line that contains '[on]' and then get the last but one
field out of that line.

Also, one or more lines may contain '[on]' in it, so we get just the
first one.