Add colors to readonly symbol in statusline
with the help of some Vimscript function to get the colors from the
highlight group parameters.

This is how it works:

We use `redir` to capture the output of `hi Special`. For example, this
currently gives me:

Special        xxx ctermfg=224 guifg=#ffa657

Out job is to get the values in the 224 and #ffa657 from this.

There is a slight complication in that some highlight groups might link
to another group. So we need to do this recursively in that case. For
instance:

DiagnosticVirtualTextError xxx links to DiagnosticError

In this case, we need to call recursively call the function with this
linked highlight group as the new parameter.