Fix blaming files that differ from version in HEAD
Previous to this, if we added an extra line to a file with x lines and
tried to blame the new line with this function, an error would occur
saying that the file only contained x lines.

This is because we need to handle differently modified files and files
that are different than the checked out revision.

We do this by using the flag `--contents`, which accept a file path. We
only do this if the file was modified (and not yet saved) or if it's
differs from the version in HEAD (we use `git diff` with `--exit-code`
flag, the exit code being 1 means it is different).