Add function to get json path using treesitter
This will compute a jq-style string to access the JSON value under the
cursor.

At root, it'll be '.'. For the first item in an array, '.[0]'. If that
item is an object with `name` keys, it'll show '.[0].name' when you pass
through it.

I'm using this function to show the path in a winbar.

I also added a mapping to copy the path to the clipboard.