Prepend accessors with dot
The issue is that for the follwing JSON, the path when the cursor in
"sha", would be `.parents.[0].sha`, which is not valid in jq, it should
be `.parents[0].sha`.

{
  "parents": [
    {
      "sha": "f2ad9517c72f6267ae317639ab56bbfd4a8653d4",
    }
  ]
}

Meaning, the logic as to whether prepend the accessor with a dot or not
is a little bit more complicated.

In this commit, the logic implemented logic is, "if the accessor
starts with [, then do not prepend a .".